Moved propagation to separated mod
