+++ crates/world_server/src/set/mod.rs
@@ -1,7 +1,9 @@
 mod chunk_initialization;
 mod chunk_management;
 mod landscape;
+mod propagation;
 
 pub use chunk_initialization::*;
 pub use chunk_management::*;
 pub use landscape::*;
+pub use propagation::*;
